Solve edge cases around content with colons
